ApiToken moved from api module to comarch-connector
