+++ fakturogen-web/src/main/java/pl/fakturogen/web/controller/rest/TokenController.java
@@ -3,7 +3,7 @@ package pl.fakturogen.web.controller.rest;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import pl.fakturogen.api.token.ApiToken;
+import pl.fakturogen.comarch.connector.model.ApiToken;
 import pl.fakturogen.comarch.connector.connectors.ComarchApiTokenConnector;
 
 import java.io.IOException;
